refactor: move interactive CLI files to correct locations and fix imports
- Move commands.ts to interactive-commands/ (not shared infrastructure)
- Move command-parser.ts to interactive-commands/ (interactive CLI only)
- Move mcp-add-utils.ts to interactive-commands/mcp/ (domain-specific)
- Fix all import paths affected by file moves
- Improve architectural clarity: interactive vs project command separation

This corrects the file organization to properly reflect that commands.ts
and command-parser.ts are specifically for interactive CLI (/commands),
not shared between project and interactive commands.
