+++ src/app/cli/interactive-commands/model/model-commands.ts
@@ -15,7 +15,7 @@
 import chalk from 'chalk';
 import { logger } from '@core/index.js';
 import { SaikiAgent } from '@core/index.js';
-import { CommandDefinition } from '../../command-parser.js';
+import { CommandDefinition } from '../command-parser.js';
 
 /**
  * Model management commands
