+++ src/app/cli/interactive-commands/session/session-commands.ts
@@ -19,7 +19,7 @@ import chalk from 'chalk';
 import { logger } from '@core/index.js';
 import { SaikiAgent } from '@core/index.js';
 import type { SessionMetadata } from '@core/index.js';
-import { CommandDefinition } from '../../command-parser.js';
+import { CommandDefinition } from '../command-parser.js';
 import { formatSessionInfo, formatHistoryMessage } from './helpers/formatters.js';
 
 /**
