+++ test-configs/no-mcp-test.yml
@@ -1,0 +1,49 @@
+# No MCP Servers Test Configuration
+# Use this for testing Saiki with only built-in capabilities (no external tools)
+
+name: "No MCP Servers Test"
+description: "Testing Saiki without any MCP servers - only built-in LLM capabilities"
+
+# System prompt for testing agent
+systemPrompt: "You are a dummy agent. You are testing with no MCP servers - only built-in LLM capabilities"
+
+# LLM Configuration (required)
+llm:
+  provider: "openai"
+  model: "gpt-4o-mini"
+  apiKey: "${OPENAI_API_KEY}"
+  maxInputTokens: 4000
+  temperature: 0.7
+
+# No MCP servers configured - this section is now optional
+# mcpServers: {}  # This is the default when omitted
+
+# Storage Configuration
+storage:
+  cache:
+    type: "memory"
+    # Optional memory storage settings:
+    # maxSize: 1000  # Maximum number of items to store
+    # ttl: 1800000   # Time-to-live in milliseconds (30 minutes)
+  database:
+    type: "memory"
+
+# Session Configuration
+session:
+  maxSessions: 5
+  sessionTimeout: 1800000  # 30 minutes in milliseconds
+  cleanupInterval: 300000  # 5 minutes in milliseconds
+
+# Logging
+logging:
+  level: "info"
+  enableConsole: true
+  enableFile: false
+
+# Test Instructions:
+# 1. Run: saiki --agent test-configs/no-mcp-test.yml
+# 2. Agent will start with only built-in LLM capabilities
+# 3. No external tools will be available
+# 4. Test basic conversation functionality
+# 5. Verify that the agent works without any MCP servers
+# 6. This demonstrates minimal Saiki configuration
